+++ web/components/user-page.tsx
@@ -40,7 +40,7 @@ import { DOMAIN, ENV_CONFIG } from 'common/envs/constants'
 import { PostCardList } from './posts/post-card'
 import { usePostsByUser } from 'web/hooks/use-post'
 import { LoadingIndicator } from './widgets/loading-indicator'
-import { DailyStats } from 'web/components/daily-stats'
+import { DailyStats } from 'web./groups/group-post-sectionats'
 import { SectionHeader } from './groups/group-about'
 import { buttonClass } from './buttons/button'
 import { MoreOptionsUserButton } from 'web/components/buttons/more-options-user-button'
