+++ web/pages/home/index.tsx
@@ -27,7 +27,7 @@ import { ContractCard } from 'web/components/contract/contract-card'
 import { ContractsGrid } from 'web/components/contract/contracts-grid'
 import { ProfitChangeTable } from 'web/components/contract/prob-change-table'
 import { DailyStats } from 'web/components/daily-stats'
-import { PinnedItems } from 'web/components/groups/group-about'
+import { PinnedItems } from 'web/components/groups/group-post-section'
 import { Col } from 'web/components/layout/col'
 import { Page } from 'web/components/layout/page'
 import { Row } from 'web/components/layout/row'
